feat: buy ticket

+++ lib/features/send/presentation/cubit/send_token_quoting_state.dart
@@ -12,7 +12,7 @@ class SendTokenQuotingInitialState extends SendTokenQuotingState {}
 class SendTokenQuotingLoadingState extends SendTokenQuotingState {}
 
 class SendTokenQuotingSuccessState extends SendTokenQuotingState {
-  final double? networkFee;
+  final int? networkFee;
   final double? exchangeRate;
   final double? amountInFiat;
